requirements.txt
streamlit>=1.28.0
plotly>=5.18.0
pyswisseph>=2.10.3
geopy>=2.4.1
pytz>=2023.3
numpy>=1.24.0
pandas>=2.1.0 